favor static configs (dev/prod); make use of webpack's 'externals' feature
