+++ src/components/actions/SaveToContact/SaveToContactForm.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { v4 as generateUUID } from 'uuid';
-import { SaveToContact, IUpdateContact } from '../../../flowTypes';
-import { Endpoints } from '../../../editor.config';
+import { Endpoints, SaveToContact, IUpdateContact } from '../../../flowTypes';
 import { Type } from '../../../services/EditorConfig';
 import { FormProps } from '../../NodeEditor';
 import ComponentMap from '../../../services/ComponentMap';
