+++ src/modules.d.ts
@@ -1,0 +1,2 @@
+/** Base config provided via Webpack's 'externals' feature */
+declare module 'Config';
