+++ src/services/EditorConfig.ts
@@ -2,7 +2,8 @@ import { ComponentClass, SFC } from 'react';
 import { AnyAction } from '../flowTypes';
 import { Language } from '../components/LanguageSelector';
 import { AnyFormProps } from '../components/NodeEditor/NodeEditor';
-import { languages, Languages, endpoints, Endpoints } from '../editor.config';
+import { Languages, Endpoints } from '../flowTypes';
+import { languages, endpoints } from 'Config';
 import ChangeGroupComp from '../components/actions/ChangeGroup/ChangeGroup';
 import ChangeGroupForm from '../components/actions/ChangeGroup/ChangeGroupForm';
 import SaveFlowResultComp from '../components/actions/SaveFlowResult/SaveFlowResult';
