+++ webpack.dev.js
@@ -1,6 +1,7 @@
 const { smartStrategy } = require('webpack-merge');
 const { HotModuleReplacementPlugin, NamedModulesPlugin } = require('webpack');
 const commonConfig = require('./webpack.common');
+const Config = require('./flowEditor.config.dev');
 
 const devConfig = {
     entry: [
@@ -65,6 +66,9 @@ const devConfig = {
                 use: ['react-hot-loader/webpack']
             }
         ]
+    },
+    externals: {
+        Config: JSON.stringify(Config)
     }
 };
